refactor(music): Use explicit barrels and .d.ts types
