+++ src/domains/music/repositories/music-repository.ts
@@ -2,11 +2,11 @@
  * @module @domains/music/repositories/music-repository
  * @remarks Database access for core music records stored in the `music` table.
  */
-import type { MusicDB } from '@domains/music/types/music-db.d-types'
 import { db } from '@db/client'
-import { eq, inArray } from 'drizzle-orm'
-import { dbError } from '@shared/errors/db-errors'
 import { music } from '@db/schemas/music'
+import type { MusicDB } from '@domains/music/types/music-db-types'
+import { dbError } from '@shared/errors/db-errors'
+import { eq, inArray } from 'drizzle-orm'
 
 /**
  * Reads music rows from the music table.
