+++ src/domains/music/types/index.d.ts
@@ -3,13 +3,13 @@
  * @remarks Barrel exports for music domain TypeScript types, including Drizzle-derived
  * database row shapes and Zod-inferred API payload types.
  * @see {@link ./music-details.d-types} for {@link MusicDetails} and nested API types
- * @see {@link ./music-db.d-types} for database row aliases
+ * @see {@link ./music-db-types} for database row aliases
  * @example
  * ```typescript
  * import type { MusicDetails, MusicDB } from '@domains/music/types'
  * ```
  */
-export * from './music-db.d-types'
-export * from './music-details.d-types'
-export * from './music-card.d-types'
-export * from './music-list.d-types'
+export * from './music-card-types'
+export * from './music-db-types'
+export * from './music-details-types'
+export * from './music-list-types'
