+++ src/domains/music/types/music-card-types.d.ts
