+++ src/domains/music/types/music-details-types.d.ts
